Refactored and polished code
