+++ pages/components/Loader.jsx
@@ -1,0 +1,19 @@
+import React from 'react'
+
+const Loader = () => {
+
+    return (
+        <div className="lds-roller">
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+        </div>   
+    )
+}
+
+export default Loader
