+++ pages/index.jsx
@@ -4,6 +4,7 @@ import { ethers } from 'ethers'
 import axios from 'axios'
 import Web3Modal from 'web3modal'
 import Image from 'next/image'
+import Loader from './components/Loader'
 
 import { nftAddress, nftMarketAddress } from '../config'
 
@@ -14,11 +15,13 @@ import DPMarket from '../artifacts/contracts/DPMarket.sol/DPMarket.json'
 const Home = () => {
 
   const [nfts, setNfts] = useState([])
-  const [loading, setLoading] = useState(true)
+  const [loading, setLoading] = useState(true)  
 
-  useEffect(() => {
-    loadNFTs()
-  }, [])
+  const renderLoading = () => (
+    <div className='loader-container'>
+        <Loader />
+    </div>
+  )
 
   const loadNFTs = async () => {
 
@@ -71,28 +74,34 @@ const Home = () => {
 
   }
 
+  useEffect(() => {
+    loadNFTs()
+  }, [])
+
   if(!loading && !nfts.length) return (<h1 className='px-20 ml-12 py-36 text-3xl text-funky-green'>No NFTs in the marketplace</h1>)
 
   return (
     <div className='p-4'>
-      <div className='px-4' style={{ maxWidth: 1600 }}>
-        <div className='grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4 pt-4 pb-4'>
-          {nfts.map((nft, i) => (
-            <div key={i} className='border shadow rounded-x1 overflow-hidden'>
-              <div style={{ position: "relative", width: "100%", height: 400 }}><Image src={nft.image} alt={nft.name} layout='fill' objectFit='cover' /></div>
-              <div className='p-4 bg-black'>
-                <p style={{ height: 64 }} className='text-3xl text-funky-green font-semibold'>{nft.name}</p>
-                <div style={{ height: 72, overflow: 'hidden' }}>
-                  <p className='text-funky-green'>{nft.description}</p>
+      <div className='px-4'>
+        {loading ? renderLoading() : 
+          <div className='grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4 pt-4 pb-4'>
+            {nfts.map((nft, i) => (
+              <div key={i} className='border shadow rounded-x1 overflow-hidden'>
+                <div style={{ position: "relative", width: "100%", height: 400 }}><Image src={nft.image} alt={nft.name} layout='fill' objectFit='cover' /></div>
+                <div className='p-4 bg-black'>
+                  <p style={{ height: 64 }} className='text-3xl text-funky-green font-semibold'>{nft.name}</p>
+                  <div style={{ height: 72, overflow: 'hidden' }}>
+                    <p className='text-funky-green'>{nft.description}</p>
+                  </div>
+                </div>
+                <div className='p-4 bg-black'>
+                  <p className='text-3x-l mb-4 font-bold text-funky-green'>{nft.price} MATIC</p>
+                  <button className='w-full bg-primary text-funky-green font-bold py-3 px-12 rounded' onClick={() => buyNFT(nft)}>Buy</button>
                 </div>
               </div>
-              <div className='p-4 bg-black'>
-                <p className='text-3x-l mb-4 font-bold text-funky-green'>{nft.price} ETH</p>
-                <button className='w-full bg-primary text-funky-green font-bold py-3 px-12 rounded' onClick={() => buyNFT(nft)}>Buy</button>
-              </div>
-            </div>
-          ))}
-        </div>
+            ))}
+          </div>
+        }
       </div>
     </div>
   )
